register_savevm etc.: make save function optional
This allows a device to provide only a loading function, not a saving
function.

This is useful if a later version of an emulation restores its state
in some other way (eg by having a different device name, or by
exploiting invariants in the state); when this happens it is still
necessary to load or skip the information when it is found in old
savefiles.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>