Update to xen-unstable 17737 plus ds->resize_shared vs. ds_resize
This one commit contains two substantial sets of changes:

 * All changes from xen-unstable from 17647 to 17737:c93a913c221f

 * Shared display buffer API revamp
    Patch from Stefano Stabellini <stefano.stabellini@eu.citrix.com>, to
    do shared display buffer stuff with a different (backward-compatible)
    API which avoids having to change all of the other display hardware
    models.  Stefano tells me this patch is likely to break stubdom in
    its current state.

The commit is the result of applying Stefano's patch to xen-unstable
hg 17737, then running  hg diff -r17647 tools/ioemu, applying the
output to the git tree with patch, and fixing up the conflicts (which
were easy).

To make future merges easier, I have included Stefano's patch rebased
to 17737 (the result of hg diff -r17737 from the above hg tree) here,
so that the base point for a future merge to a new xen-unstable tip is
easy to recreate.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>