New scriptarg=... -net parameter allows passing of an argument to the script.
Previously, network scripts would have to do all of their work based
only on the script name and interface name.  If the script's behaviour
is supposed to be different for different network interfaces this
might involve constructing a special script for each interface with
the associated need to delete it etc.

With this patch it is possible to specify
  -net=...,scriptarg=<extra-info>
which gets passed as a $2 to the qemu-ifup script.