Store pty name in xenstore for the benefit of the tools.
This is a rather intrusive patch, because we need a mechanism to get
the ptsname out of the guts of the qemu startup code.  This is a
first cut of such a thing which might eventually go upstream.

The main new interface is chr_getname, a new CharDriverState method.
If implemented it returns "<type> <name>" where currently "<type>" is
"pty" for ptys.  If not implemented the function pointer may be null.

We import store_dev_info from xen-unstable (17987:9b35ae586cb8) and
eviscerate it appropriately.