Nicer fix for spurious messages from configure when GCC3 is missing.
This reverts b5123e05cce4be4c6c8c822fad0f0df4c053da06 and replaces it
with a nice simple change, namely to change  test ... -a ...  with
test ... && test ... .  That provides shortcut evaluation.

Thanks to Andre Przywara who provided this one-line change as an
update to the previous commit.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>