vga.c: Call dpy_update when stealing vram
When we steal the vram, we copy the old data into it.  It may be the
case that the display output is already running - eg a VNC client
already connected.  In this case it is necessary to call dpy_update to
ensure that the data is sent to the client.

Without this change the initial screen clearing (done by the BIOS
before vram is stole) may be missed if the vnc client is fast at VM
startup, resulting in a white vnc screen with bits of black-backed
text painted over (which can be rectified by asking for a screen
refresh).

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>