posix_aio_init: Explicitly unblock SIGUSR1
On Centos 4.3 as invoked by xend it seems that SIGUSR1 can be blocked.

This is almost certainly a libc bug; I checked by using sigprocmask to
obtain what was allegedly the current signal mask and printing it out
and SIGUSR1 wasn't listed, although my strace showed rt_sigprocmask
calls which clearly implied it was blocked - and indeed it wasn't
delivered.

Explicitly enabling it is an easy workaround.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>