xen-hooks.mak: Do not print `***' when compiling unless we fall over
make prints `***' when it stops due to an error.  With make -j, and/or
with automated build systems, this error marker can be in the middle
of a large logfile.  It is useful to be able to search for it.

Therefore change occurrences of `*' to `#' for information and warning
messages.  `***' is reserved for cases where the build stops.

A corresponding change is being submitted to xen-unstable.hg.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>