fix xenfb frontend and backend
Due to the DisplayState changes the initialization order is changed
as well.

Now graphic devices are the first to be initialized and only after them
any display frontend is initialized.

Obviously this change has a direct consequence on the xenfb backend and
frontend and this patch takes care of adapting them to the new
initialization order.

Signed-off-by: Stefano Stabellini <stefano.stabellini@eu.citrix.com>