qemu xen upstream synch (Gerd): [PATCH 3/7]: fix fsf address & compare style
changes coming from upstream merge review (style only, no code changes).
 - Update FSF address.
 - Fix strage compare style.

[This is] a series of patches for qemu-xen, making the code identical
to the xen support patches being submitted to upstream qemu.  The review
process on qemu-devel resulted in a number of fixes and cleanups in the
backend code, this is where most of the changes come from.

There are also some xenfb changes due to displaystate reorganization and
xenfb being merged in steps due to that.

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>