replace stat("/etc/disable-guest-log-throttle") in xen_platform.c
currently xen_platform reads /etc/disable-guest-log-throttle at boot to
decide whether is going to throttle guest logging or not.
When using stubdom this is the only code in qemu that rely on
fs-backend, apart from saving and loading the statefile for migration.
This patch removes the read from /etc and adds a read from xenstore
instead.

Signed-off-by: Stefano Stabellini <stefano.stabellini@eu.citrix.com>