Allow changing CD for /dev/xvdX devices.
We found the issue being not able to change CD on the HVM-Domain.
It is possible on the /dev/hdc device,
but it is impossible on the/dev/xvdc device.

We want to work it as all /dev/xvdX devices on the HVM-domain
 as well as on the PV-domain.

Signed-off-by: Takanori Kasai <kasai.takanori@jp.fujitsu.com>

The execution method is as follows.
----------------------------------------------------------------------
Domain configuration file:
 disk = ["tap:aio:/<guest image file>,xvda,w", ",xvdc:cdrom,r"]

Operation that assign CD:
 # xm block-configure <domain> file:<iso image> xvdc:cdrom r

Operation that releases CD
 # xm block-configure <domain> '' xvdc:cdrom r
----------------------------------------------------------------------