passthrough: enabling IGD passthrough for Calpella and Sandybridge
This patch enables IGD passthrough for Calpella and Sandybridge
platforms.  To minimize impact of these changes, it checks for
vendor_id of 0x8086 before creating another PCH device in the virtual
platform.  For opregion, it checks both vendor_ID of 0x8086 and a
non-zero PCI opregion value on device 0:2.0 before mapping the
opregion.

Signed-off-by: Allen Kay <allen.m.kay@intel.com>