xen: introduce an event channel for buffered io event notifications
Use the newly [November 2011] introduced HVM_PARAM_BUFIOREQ_EVTCHN to
receive notifications for buffered io events.  After the first
notification is received leave the event channel masked and setup a
timer to process the rest of the batch.  Once we have completed
processing the batch, unmask the event channel and delete the timer.

Signed-off-by: Stefano Stabellini <stefano.stabellini@eu.citrix.com>